requirements.txt
appdirs==1.4.3
click==6.7
Flask==0.12
Flask-Login==0.4.0
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==1.0
MySQL-python==1.2.5
packaging==16.8
pkg-resources==0.0.0
pyparsing==2.2.0
python-dateutil==2.6.0
six==1.10.0
Werkzeug==0.11.15